Rename Attention_24_23.h to attention_24_23.h for consistent naming
All other adapter headers use lowercase filenames; align this one to match.
