+++ onnx/version_converter/convert.h
@@ -17,7 +17,7 @@
 #include <vector>
 
 #include "onnx/version_converter/BaseConverter.h"
-#include "onnx/version_converter/adapters/Attention_24_23.h"
+#include "onnx/version_converter/adapters/attention_24_23.h"
 #include "onnx/version_converter/adapters/axes_attribute_to_input.h"
 #include "onnx/version_converter/adapters/axes_input_to_attribute.h"
 #include "onnx/version_converter/adapters/axis_attribute_to_input.h"
